% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{convertDFtoScoreMatrix}
\alias{convertDFtoScoreMatrix}
\title{Convert an MRMC data frame to a score matrix, dropping readers or cases with no observations}
\usage{
convertDFtoScoreMatrix(dfMRMC)
}
\arguments{
\item{dfMRMC}{An MRMC data frame}
}
\value{
scores
}
\description{
Convert an MRMC data frame to a score matrix, dropping readers or cases with no observations
}
