% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createGroups}
\alias{createGroups}
\title{Assign a group label to items in a vector}
\usage{
createGroups(items, nG)
}
\arguments{
\item{items}{A vector of items}

\item{nG}{The number of groups}
}
\value{
A data frame containing the items and their group labels
}
\description{
Assign a group label to items in a vector
}
\examples{
x <- paste("item", 1:10, sep = "")
df <- createGroups(x, 3)
print(df)

}
