% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMRMC.R
\name{createIMRMCdf}
\alias{createIMRMCdf}
\title{Title}
\usage{
createIMRMCdf(dFrame, dataColumnNames, truePositiveFactor)
}
\arguments{
\item{dFrame}{This data frame includes columns for readerID, caseID, modalityID, score, and truth.
These columns are not expected to be named as such and other columns may exist.}

\item{dataColumnNames}{This list identifies the column names of the data frame to be used for the analysis.
list(readerID = "***", caseID = "***", modalityID = "***", score = "***", truth="***")}

\item{truePositiveFactor}{The true positive label, such as "cancer" or "1"}
}
\value{

}
\description{
Title
}
