% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{deleteCol}
\alias{deleteCol}
\title{Delete a data frame column}
\usage{
deleteCol(df, colName)
}
\arguments{
\item{df}{A data frame}

\item{colName}{Column name or list of column names to be deleted}
}
\value{
The data frame without the deleted column or columns
}
\description{
Delete a data frame column
}
