% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doROCcurveMRMC.R
\name{doROCcurveMRMC}
\alias{doROCcurveMRMC}
\title{Create a standard set of ROC curves from an MRMC data frame}
\usage{
doROCcurveMRMC(mrmcAlternate)
}
\arguments{
\item{mrmcAlternate}{data frame
\itemize{
  \item{\code{readerID}}
  \item{\code{caseID}}
  \item{\code{modalityID}}
  \item{\code{score}}
  \item{\code{truthLabel}}
}}
}
\value{
list of ROC curves corresponding to
all reader x modality combinations,
all modalities with the reader scores pooled,
all modalities with the per-reader ROC curves, horizontally, diagonally, 
and vertically averaged.
}
\description{
Create a standard set of ROC curves from an MRMC data frame
}
