% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doROCcurveMRMC.R
\name{doROCxyMRMC}
\alias{doROCxyMRMC}
\title{Create empirical ROC curve from an MRMC formatted data frame}
\usage{
doROCxyMRMC(mrmcAlternate)
}
\arguments{
\item{mrmcAlternate}{data frame
\itemize{
  \item{\code{readerID}}
  \item{\code{caseID}}
  \item{\code{modalityID}}
  \item{\code{score}}
  \item{\code{truthLabel}}
}}
}
\value{
data frame of an ROC curve
\itemize{
  \item{\code{modalityID}}
  \item{\code{readerID}}
  \item{\code{fpf} False-positive fractions (1-specificity)}
  \item{\code{tpf} True-positive fractions (sensitivity)}
  \item{\code{threshold} Threshold corresponding to each fpf, tpf}
}
}
\description{
Create empirical ROC curve from an MRMC formatted data frame
}
