% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractPairedComparisonsBRBM}
\alias{extractPairedComparisonsBRBM}
\title{Extract between-reader between-modality pairs of scores}
\usage{
extractPairedComparisonsBRBM(
  data0,
  modalities = c("testA", "testB"),
  keyColumns = list(readerID = "readerID", caseID = "caseID", modalityID = "modalityID",
    score = "score")
)
}
\arguments{
\item{data0}{This data frame includes columns for readerID, caseID, modalityID, score.}

\item{modalities}{The modalities (testA, testB) for the scores to be paired}

\item{keyColumns}{This list identifies the column names
of the data frame to be used for the analysis.
       list(readerID = "***", caseID = "***",
            modalityID = "***", score = "***", truth="***")}
}
\value{
A data frame of all paired observations.
  Each observation comes from a pair of readers evaluating a case in two modalities.
  The first column corresponds to one reader evaluating the case in testA.
  The second column corresonds to the other reader evaluating the case in testB.
}
\description{
Extract between-reader between-modality pairs of scores
}
