% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitsOfAgreement.R
\name{getBRBM}
\alias{getBRBM}
\title{Get between-reader, between-modality paired data from an MRMC data file}
\usage{
getBRBM(mcsData, modality.X, modality.Y)
}
\arguments{
\item{mcsData}{A data frame with the following columns: readerID, caseID, modalityID, score}

\item{modality.X}{The name of one modality}

\item{modality.Y}{The name of one modality.}
}
\value{
The result of merging the modality.X and modality.Y subsets of mcsData by caseID
 for every pair of readers
}
\description{
Get between-reader, between-modality paired data from an MRMC data file
}
\details{
If modality.Y = modality.X, then the data would be between-reader, within-modality (BRWM).
}
