% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMRMCscore}
\alias{getMRMCscore}
\title{Get a score from an MRMC data frame}
\usage{
getMRMCscore(df, iR, iC, modality)
}
\arguments{
\item{df}{An MRMC data frame}

\item{iR}{The numeric index of the readerID}

\item{iC}{The numeric index of the caseID}

\item{modality}{The character description of the modalityID}
}
\value{
The score
}
\description{
Get a score from an MRMC data frame
}
