% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{init.lecuyerRNG}
\alias{init.lecuyerRNG}
\title{Initialize the l'Ecuyer random number generator}
\usage{
init.lecuyerRNG(seed = 1, stream = 2)
}
\arguments{
\item{seed}{This determines the position in each stream}

\item{stream}{This determines the stream}
}
\value{
Nothing
}
\description{
See the documentation for the parallel package.
If you require backwards compatibility, please run \code{RNGversion("3.5.0")}.
}
