% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericalMRMC.R
\name{meanModelA}
\alias{meanModelA}
\title{Numerically calculate the mean of model A. GLMM with
  fixed mean
  reader and case cross-correlated effects
  a reader by case interaction term}
\usage{
meanModelA(simModelAconfig, showPlots = FALSE, nSamples = 400)
}
\arguments{
\item{simModelAconfig}{list of simulation parameters:
Nr         number of readers
Nc         number of cases
mu         fixed effect: mean in logistic space
var_c      variance of random case effect
var_r      variance of random reader effect
var_rc     variance of randome reader by case effect}

\item{showPlots}{= FALSE (default) If TRUE, function shows plots of integrand
in probability space and logistic space.}

\item{nSamples}{= 400 (default) controls precision of integration}
}
\value{
Mean of model A by numerical integration
}
\description{
Numerically calculate the mean of model A. GLMM with
  fixed mean
  reader and case cross-correlated effects
  a reader by case interaction term
}
