% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renameCol}
\alias{renameCol}
\title{Rename a data frame column name or a list object name}
\usage{
renameCol(df, oldColName, newColName)
}
\arguments{
\item{df}{A data frame}

\item{oldColName}{Old column name}

\item{newColName}{New column name}
}
\value{
the data frame with the updated column name
}
\description{
Rename a data frame column name or a list object name
}
