% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roc2binary}
\alias{roc2binary}
\title{Convert ROC data formatted for doIMRMC to TPF and FPF data formatted for doIMRMC}
\usage{
roc2binary(df.auc, threshold)
}
\arguments{
\item{df.auc}{data frame of roc scores formatted for doIMRMC}

\item{threshold}{The threshold for determining binary decisions}
}
\value{
a list of two data frames (df.tpf and df.fpf) both formatted for doIMRMC
}
\description{
Convert ROC data formatted for doIMRMC to TPF and FPF data formatted for doIMRMC
}
