% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{sim.gRoeMetz.config}
\alias{sim.gRoeMetz.config}
\title{sim.gRoeMetz.config creates a configuration file for the sim.gRoeMetz program}
\usage{
sim.gRoeMetz.config(nR = 10, nC.neg = 100, nC.pos = 100, mu.neg = 0,
  mu.pos = 1, var_r = 0.03, var_c = 0.3, var_rc = 0.2)
}
\arguments{
\item{nR}{Number of readers (default = 10)}

\item{nC.neg}{Number of signal-absent cases (default = 100)}

\item{nC.pos}{Number of signal-present cases (default = 100)}

\item{mu.neg}{Mean fixed effect of signal-absent distribution (default = 0.0) \cr
Modality specific parameters are set to zero: mu.Aneg = mu.Bneg = 0}

\item{mu.pos}{Mean fixed effect of signal-present distribution (default = 1.0) \cr
Modality specific parameters are set to zero: mu.Apos = mu.Bpos = 0}

\item{var_r}{Variance of reader random effect (default = 0.03) \cr
var_r.neg = var_r.pos = var_r.Aneg = var_r.Apos = var_r.Bneg = var_r.Bpos = var_r \cr}

\item{var_c}{Variance of case random effect (default = 0.30) \cr
var_c.neg = var_c.pos = var_c.Aneg = var_c.Apos = var_c.Bneg = var_c.Bpos = var_c \cr}

\item{var_rc}{Variance of reader.by.case random effect (default = 0.20) \cr
var_rc.neg = var_rc.pos = var_rc.Aneg = var_rc.Apos = var_rc.Bneg = var_rc.Bpos = var_rc \cr}
}
\value{
config [list] Refer to the simRoeMetz input variable
}
\description{
This function creates a configuration file for the generalized Roe & Metz
simulation model as described in Gallas2014_J-Med-Img_v1p031006.
The default model returned when there are no arguments given to the function is
the HH model from Roe1887_Acad-Radiol_v4p298. Following that paper,
The user can also specify three parameters related to experiment size (nR, nC.neg, nC.pos)
and five parameters parameters specifying a linear model that do not
depend on modality or truth (mu.neg, mu.pos, var_r, var_c, var_rc).
}
\details{
If no arguements, this function returns a default simulation configuration for sim.gRoeMetz
}
\examples{
config <- sim.gRoeMetz.config()
}
