% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{sim.gRoeMetz}
\alias{sim.gRoeMetz}
\title{simRoeMetz Simulate an MRMC data set for an ROC experiment comparing two modalities
  references: Roe1997_Acad-Radiol_v4p298, Roe1997_Acad-Radiol_v4p587
  references: Gallas2014_J-Med-Img_v1p031006}
\usage{
sim.gRoeMetz(config)
}
\arguments{
\item{config}{[list] of simulation parameters:
         modalityID.A     [chr] label modality A
         nR               [num] number of readers
         nC.neg            [num] number of signal-absent cases
         nC.pos            [num] number of signal-present cases

         The model parameters include for six fixed effects:
           mu.neg = signal-absent (neg, global mean)
           mu.pos = signal-present (pos, global mean)
           mu.Aneg = modality A signal-absent (Aneg, modality effect)
           mu.Bneg = modality B signal-absent (Bneg, modality effect)
           mu.Apos = modality A signal-present (Apos, modality effect)
           mu.Bpos = modality B signal-present (Bpos, modality effect)

         The model parameters also include 18 random effects.
         Six that are independent of modality
           var_r.neg         [num] variance of random reader effect
           var_c.neg         [num] variance of random case effect
           var_rc.neg        [num] variance of randome reader by case effect
           var_r.pos         [num] variance of random reader effect
           var_c.pos         [num] variance of random case effect
           var_rc.pos        [num] variance of randome reader by case effect
         Six that are specific to modality A
           var_r.Aneg         [num] variance of random reader effect
           var_c.Aneg         [num] variance of random case effect
           var_rc.Aneg        [num] variance of randome reader by case effect
           var_r.Apos         [num] variance of random reader effect
           var_c.Apos         [num] variance of random case effect
           var_rc.Apos        [num] variance of randome reader by case effect
         Six that are specific to modality B
           var_r.Bneg         [num] variance of random reader effect
           var_c.Bneg         [num] variance of random case effect
           var_rc.Bneg        [num] variance of randome reader by case effect
           var_r.Bpos         [num] variance of random reader effect
           var_c.Bpos         [num] variance of random case effect
           var_rc.Bpos        [num] variance of randome reader by case effect}
}
\value{
dFrame.imrmc   [data.frame] with (nC.neg + nC.pos)*nR rows including
           readerID       [Factor] w/ nR levels "reader1", "reader2", ...
           caseID         [Factor] w/ nC levels "case1", "case2", ...
           modalityID     [Factor] w/ 1 level config$modalityID
           score          [num] reader score
}
\description{
linear model
  L.mrct = mu.t + mu.mt + reader.rt + case.ct + readerXcase.rct
                 + modalityXreader.mrt + modalityXcase.mct + modalityXreaderXcase.mrct
    mu.t treats is the global fixed
    mu.mt treats the modality-specific fixed effects
      m=modality (levels: A and b)
      t=truth (levels: neg and Pos)
    the remaining terms are the random effects: all independent normal random variables
}
\examples{
config <- sim.gRoeMetz.config()
simRoeMetz.imrmc <- sim.gRoeMetz(config)

}
