% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{simMRMC}
\alias{simMRMC}
\title{simMRMC Simulate an MRMC data set (Roe1997_Acad-Radiol_v4p298, Roe1997_Acad-Radiol_v4p587)
  linear model with independent random effects that are normally distributed
  L.rc = mu + readerEffect.r + caseEffect.c + readerXcaseEffect.rc}
\usage{
simMRMC(simMRMC.config)
}
\arguments{
\item{simMRMC.config}{[list] of simulation parameters:
modalityID [chr] label modalityID
readerIDs  [factor] the ID of each reader
caseIDs    [factor] the ID of each case
mu         [num] mean
var_r      [num] variance of random reader effect
var_c      [num] variance of random case effect
var_rc     [num] variance of random reader by case effect}
}
\value{
L [data.frame] with nC*nR rows of 4 variables
           L$modalityID   [factor] determined by input modalityID
           L$readerID     [factor] determined by input readerIDs
           L$caseID       [factor] determined by input caseIDs
           L$score        [num]  R.r + C.c + RC.rc
                            r = 1,2,...,nR
                            c = 1,2,...,nC
                            R.r ~ N(0,var_r)
                            C.c ~ N(0,var_c)
                            RC.rc ~ N(0,var_rc)
}
\description{
simMRMC Simulate an MRMC data set (Roe1997_Acad-Radiol_v4p298, Roe1997_Acad-Radiol_v4p587)
  linear model with independent random effects that are normally distributed
  L.rc = mu + readerEffect.r + caseEffect.c + readerXcaseEffect.rc
}
