% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{simRoeMetz.defaultConfig}
\alias{simRoeMetz.defaultConfig}
\title{simRoeMetz.defaultConfig creates a default configuration file for the simRoeMetz program}
\usage{
simRoeMetz.defaultConfig()
}
\value{
simRoeMetz.config [list] Refer to the simRoeMetz input variable
}
\description{
simRoeMetz.defaultConfig creates a default configuration file for the simRoeMetz program
}
\examples{
simRoeMetz.config <- simRoeMetz.defaultConfig()
}
