% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{simRoeMetz.example}
\alias{simRoeMetz.example}
\title{Simulates a sample MRMC ROC experiment}
\usage{
simRoeMetz.example()
}
\value{
dFrame.imrmc [data.frame] Please refer to the description of the simRoeMetz return variable
}
\description{
Simulates a sample MRMC ROC experiment
}
