% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{simRoeMetz}
\alias{simRoeMetz}
\title{simRoeMetz Simulate an MRMC data set for an ROC experiment comparing two modalities
  references: Roe1997_Acad-Radiol_v4p298, Roe1997_Acad-Radiol_v4p587
  references: Gallas2014_J-Med-Img_v1p031006}
\usage{
simRoeMetz(simRoeMetz.config)
}
\arguments{
\item{simRoeMetz.config}{[list] of simulation parameters:
         modalityID.A     [chr] label modality A
         nR               [num] number of readers
         nC.neg            [num] number of signal-absent cases
         nC.pos            [num] number of signal-present cases

         There are also model parameters for six random effects:
           signal-absent (neg, global mean)
           signal-present (pos, global mean)
           modality A signal-absent (Aneg, modality effect)
           modality B signal-absent (Bneg, modality effect)
           modality A signal-present (Apos, modality effect)
           modality B signal-present (Bpos, modality effect)

         Each random effect carries its own model parameters
         Here are the parameters for the signal-absent cases (not modality specific part)
           mu.neg            [num] mean
           var_r.neg         [num] variance of random reader effect
           var_c.neg         [num] variance of random case effect
           var_rc.neg        [num] variance of randome reader by case effect}
}
\value{
dFrame.imrmc   [data.frame] with (nC.neg + nC.pos)*nR rows including
           readerID       [Factor] w/ nR levels "reader1", "reader2", ...
           caseID         [Factor] w/ nC levels "case1", "case2", ...
           modalityID     [Factor] w/ 1 level simRoeMetz.config$modalityID
           score          [num] reader score
}
\description{
linear model
  L.mrct = mu.t + mu.mt + reader.rt + case.ct + readerXcase.rct
                 + modalityXreader.mrt + modalityXcase.mct + modalityXreaderXcase.mrct
    mu.t treats is the global fixed
    mu.mt treats the modality-specific fixed effects
      m=modality (levels: A and b)
      t=truth (levels: neg and Pos)
    the remaining terms are the random effects: all independent normal random variables
}
\examples{
simRoeMetz.config <- simRoeMetz.defaultConfig()
simRoeMetz.imrmc <- simRoeMetz(simRoeMetz.config)

}
