% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{successDFtoROCdf}
\alias{successDFtoROCdf}
\title{successDFtoROCdf}
\usage{
successDFtoROCdf(df)
}
\arguments{
\item{df}{Each row contains a success observation for one reader evaluating one case}
}
\value{
data frame ready for doIMRMC
}
\description{
successDFtoROCdf
}
