% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMRMC.R
\name{undoIMRMCdf}
\alias{undoIMRMCdf}
\title{Title}
\usage{
undoIMRMCdf(df.MRMC)
}
\arguments{
\item{df.MRMC}{This data frame includes columns for readerID, caseID, modalityID, score.
Each row is a reader x case x modality observation from the study
In addition to observations from the study,
    this data frame requires rows specifying the truth for each caseID.
For truth specifications, the readerID needs to equal "truth" or "-1",
    modalityID can be anything ("truth" is a good choice),
    and score should be 0 for signal-absent normal case, 1 for signal-present disease case.}
}
\value{
output
}
\description{
Title
}
