% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{convertDFtoDesignMatrix}
\alias{convertDFtoDesignMatrix}
\title{Convert an MRMC data frame to a design matrix}
\usage{
convertDFtoDesignMatrix(dfMRMC)
}
\arguments{
\item{dfMRMC}{An MRMC data frame}
}
\value{
A matrix [nCases, nReaders] indicating which scores were reported for each reader and case
}
\description{
Convert an MRMC data frame to a design matrix, dropping readers or cases with no observations
}
