% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createIMRMCdf}
\alias{createIMRMCdf}
\title{Convert a data frame with all needed factors to doIMRMC formatted data frame}
\usage{
createIMRMCdf(
  dFrame,
  keyColumns = list(readerID = "readerID", caseID = "caseID", modalityID = "modalityID",
    score = "score", truth = "truth"),
  truePositiveFactor = "cancer"
)
}
\arguments{
\item{dFrame}{This data frame includes columns for readerID, caseID, modalityID, score, and truth.
These columns are not expected to be named as such and other columns may exist.}

\item{keyColumns}{This list identifies the column names of the data frame to be used for the analysis.
list(readerID = "***", caseID = "***", modalityID = "***", score = "***", truth="***")}

\item{truePositiveFactor}{The true positive label, such as "cancer" or "1"}
}
\value{
output a doIMRMC formatted data frame: rows for truth and rows for data
}
\description{
Convert a data frame with all needed factors to doIMRMC formatted data frame
}
