% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitsOfAgreement.R
\name{laWRBM}
\alias{laWRBM}
\title{MRMC analysis of within-reader between-modality limits of agreement}
\usage{
laWRBM(
  df,
  modalitiesToCompare,
  keyColumns = c("readerID", "caseID", "modalityID", "score")
)
}
\arguments{
\item{df}{an iMRMC formatted data frame, see \link{dfMRMC_example}}

\item{modalitiesToCompare}{The factors identifying the modalities to compare.}

\item{keyColumns}{Identify the factors corresponding to the readerID, caseID, modalityID, and score
(or alternative random and fixed effects).}
}
\description{
The core analysis is done by ustat11 with the difference kernel (kernelFlag = 2).
  This calculation can also be accomplished by ustat11 with the identity kernel (kernelFlag = 1),
  and the code to do that is provided after the return statement so it never gets executed.
}
