% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRMC.R
\name{simRoeMetz.example}
\alias{simRoeMetz.example}
\title{Simulates a sample MRMC ROC experiment}
\usage{
simRoeMetz.example()
}
\value{
dFrame.imrmc [data.frame] Please refer to the description of the simRoeMetz return variable
}
\description{
Simulates a sample MRMC ROC experiment
}
\examples{
# Simulate a sample MRMC ROC data set
dFrame.imrmc <- simRoeMetz.example()
# Analyze the MRMC ROC data
result <- doIMRMC(dFrame.imrmc)

}
