% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uStats.R
\name{uStat11.diff}
\alias{uStat11.diff}
\title{Create the kernel and design matrices for uStat11}
\usage{
uStat11.diff(df.input, keyColumns = c("readerID", "caseID", "modalityID",
  "score"), modalitiesToCompare = c("modalityA", "modalityB", "modalityA",
  "modalityB"))
}
\arguments{
\item{df.input}{Data frame of observations, one per row. Columns also identify random and fixed effects.}

\item{keyColumns}{The required columns}

\item{modalitiesToCompare}{The factors identifying the modalities to compare}
}
\description{
The kernel is the difference kernel
}
